Move handle_evid to separate file
