+++ man/reexports.Rd
@@ -38,3 +38,4 @@ below to see their documentation.
   \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
 }}
 
+\value{ Inherited from parent routine }
