+++ src/Makevars.in
@@ -16,11 +16,11 @@ PKG_CFLAGS  =  -D_isRxODE_
 PKG_CXXFLAGS = $(SHLIB_OPENMP_CXXFLAGS) -D_isRxODE_ -DBOOST_DISABLE_ASSERTS -DBOOST_NO_CXX11_STATIC_ASSERT -DBOOST_PHOENIX_NO_VARIADIC_EXPRESSION -DBOOST_NO_AUTO_PTR -D_REENTRANT -@ISYSTEM@"$(BH)" -@ISYSTEM@"$(EG)" $(SH) -@ISYSTEM@"$(RCPP)" -@ISYSTEM@"$(ARMA)"
 
 
-SOURCES_C = strncmp.c checkmate_stub.c print_node.c genModelVars.c codegen.c getOption.c sbuf.c parseLinCmt.c tran.c lincmt.c utilc.c call_dvode.c dop853.c omegaChol.c init.c cfode.c common.c corfailure.c correction.c daxpy.c ddot.c dgefa.c dgesl.c dscal.c fnorm.c idamax.c intdy.c lsoda.c methodswitch.c orderswitch.c prja.c scaleh.c solsy.c stoda.c vmnorm.c strdup_printf.c rprintf.c box.c matexp_HM98.c 
+SOURCES_C = handle_evid.c strncmp.c checkmate_stub.c print_node.c genModelVars.c codegen.c getOption.c sbuf.c parseLinCmt.c tran.c lincmt.c utilc.c call_dvode.c dop853.c omegaChol.c init.c cfode.c common.c corfailure.c correction.c daxpy.c ddot.c dgefa.c dgesl.c dscal.c fnorm.c idamax.c intdy.c lsoda.c methodswitch.c orderswitch.c prja.c scaleh.c solsy.c stoda.c vmnorm.c strdup_printf.c rprintf.c box.c matexp_HM98.c 
 SOURCES_CPP = rxvandercorput.cpp seed.cpp rxthreefry.cpp cbindThetaOmega.cpp cvPost.cpp checkmate.cpp utilcpp.cpp qs.cpp rxData.cpp lincmtB.cpp par_solve.cpp etTran.cpp forder.cpp et.cpp boost.cpp fast_factor.cpp expandGrid.cpp  RcppExports.cpp WinDrive.cpp rxInv.cpp rxStack.cpp rxOptRep.cpp expm.cpp 
 SOURCES_FORTAN = dlsoda.f opkda2.f matexp.f dsphiv.f dgpadm.f
 
-OBJECTS = $(SOURCES_CPP:.cpp=.o) $(SOURCES_C:.c=.o) $(SOURCES_FORTAN:.f=.o)
+OBJECTS = $(SOURCES_C:.c=.o) $(SOURCES_CPP:.cpp=.o) $(SOURCES_FORTAN:.f=.o)
 
 .PHONY: all 
 all: $(SHLIB)
